# ПИС/Lab11/docs/swagger.yaml
basePath: /
definitions:
  main.Celebrity:
    description: Модель данных знаменитости
    properties:
      fullName:
        example: John Doe
        type: string
      id:
        example: 1
        type: integer
      nationality:
        example: American
        type: string
      reqPhotoPath:
        example: /photos/john_doe.jpg
        type: string
    type: object
host: localhost:3000
info:
  contact: {}
  description: RESTful API веб-сервер для работы с реляционной БД (CRUD)
  title: Celebrity CRUD API (GO11_01)
  version: "1.0"
paths:
  /Celebrities:
    post:
      consumes:
      - application/json
      description: Создает новую запись в базе данных
      parameters:
      - description: Данные знаменитости (Id игнорируется)
        in: body
        name: celebrity
        required: true
        schema:
          $ref: '#/definitions/main.Celebrity'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.Celebrity'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Добавить новую знаменитость
      tags:
      - Celebrities
  /Celebrities/{id}:
    delete:
      description: Удаляет запись из базы данных по ID
      parameters:
      - description: ID знаменитости
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            type: string
        "404":
          description: Not found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Удалить знаменитость
      tags:
      - Celebrities
    get:
      description: Возвращает конкретную запись по её идентификатору
      parameters:
      - description: ID знаменитости
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Celebrity'
        "404":
          description: Not found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Получить знаменитость по ID
      tags:
      - Celebrities
    put:
      consumes:
      - application/json
      description: Полностью обновляет существующую запись по ID
      parameters:
      - description: ID знаменитости
        in: path
        name: id
        required: true
        type: integer
      - description: Новые данные
        in: body
        name: celebrity
        required: true
        schema:
          $ref: '#/definitions/main.Celebrity'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Celebrity'
        "404":
          description: Not found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Обновить данные знаменитости
      tags:
      - Celebrities
  /Celebrities/All:
    get:
      description: Возвращает список всех записей из базы данных
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Celebrity'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Получить всех знаменитостей
      tags:
      - Celebrities
swagger: "2.0"